refactor complete code

+++ services/user.service.js
@@ -1,27 +1,22 @@
 const bcrypt = require('bcrypt');
-const { hash } = require('bcrypt');
 const jwt = require('jsonwebtoken');
 const { Op } = require('sequelize');
-
 const models = require('../models');
 const { sequelize } = require('../models');
 const mailer = require('../helper/sendmail');
 const { adminAddReportee } = require('./userReportee.service');
 const redisClient = require('../utility/redis');
+const UniqueStringGenerator = require('unique-string-generator');
 
-const resetPassword = async (newPassword, id, userEmail) => {
-  
-  if (id) {
-    await models.User.update({ password: await hash(newPassword,10) }, { where: { id: id } });
-  }
-  else {
-    await models.User.update({ password: await hash(newPassword, 10) }, { where: { email: userEmail } });
-  }
+const resetPassword = async (newPassword,userEmail) => {  
+    await models.User.update({ password: await bcrypt.hash(newPassword, 10) }, { where: { email: userEmail } });
     const email_body = `Password reset successfull`;
     const email_subject = `Password reset`;
     await mailer.sendMail(email_body, email_subject, userEmail);
     return "Password reset successfully";
 }
+
+
 
 const loginUser = async (payload) => {
   const { email, password } = payload;
@@ -29,30 +24,41 @@ const loginUser = async (payload) => {
   const user = await models.User.findOne({
     where: {
       email: email
-    }
+    },
+    include: [{
+      model: models.Designation,
+      attributes: ["designation_title"]
+    },
+    {
+      model: models.Role,
+      attributes: ["role_title"]
+    },
+    {
+      model: models.User,
+      as: 'reportee_of',
+      attributes: { exclude: ["password", "created_at", "updated_at", "deleted_at", "UserReporteeMapping"] }
+    }],
+    attributes: { exclude: ["created_at", "updated_at", "deleted_at"] }
   });
   if (!user) {
-    throw new Error('Credentials are invalid!');
+    throw new Error('User Not Found!');
   }
 
-
-  const match = await bcrypt.compareSync(password, user.password);
+  const match = await bcrypt.compareSync(password, user.dataValues.password);
   if (!match) {
     throw new Error('Wrong email or password');
   }
 
-  // jwt token assignment
   const accessToken = jwt.sign({ userId: user.id }, process.env.secretKey, {
     expiresIn: process.env.jwtExpiration
   });
-  let refreshToken = await models.RefreshToken.createToken(user);
+  const refreshToken = jwt.sign({ userId: user.id }, process.env.secretKey, {
+    expiresIn: process.env.jwtRefreshExpiration
+  });
 
-  const data = {
-    refresh_token: refreshToken,
-    user_id: user.id,
-    email: user.email
-  };
-  await redisClient.set("refresh_token_detail", JSON.stringify(data));
+  delete user.dataValues.password;
+  await redisClient.set(user.id, JSON.stringify(user));
+  await redisClient.set(refreshToken, JSON.stringify("true"));
 
   return {
     id: user.id,
@@ -63,24 +69,21 @@ const loginUser = async (payload) => {
 }
 
 const refreshToken = async (requestToken) => {
-  if (!requestToken) throw new Error('Refresh Token is required!');
-
-  let refreshToken = await redisClient.get("refresh_token_detail");
-  refreshToken = JSON.parse(refreshToken);
+  let refreshToken = await redisClient.get(requestToken);
 
   if (!refreshToken) {
-    throw new Error('Refresh token is not in database!');
+    throw new Error('Login Required');
   }
-  if (refreshToken.refresh_token === requestToken) console.log("hello");
 
-  const userId = refreshToken.user_id;
-  let newAccessToken = jwt.sign({ userId: userId }, process.env.secretKey, {
+  const decoded_jwt = jwt.verify(requestToken, process.env.secretKey);
+  
+  let newAccessToken = jwt.sign({ userId: decoded_jwt.userId }, process.env.secretKey, {
     expiresIn: process.env.jwtExpiration
   });
 
   return {
     accessToken: newAccessToken,
-    refreshToken: refreshToken.refresh_token,
+    refreshToken:requestToken,
   }
 }
 
@@ -96,75 +99,53 @@ const getAllUsers = async () => {
 
 const logoutUser = async (requestToken) => {
   await redisClient.del("refresh_token_detail");
-  
   return;
-
 }
 
 const resetUserPassword = async (payload) => {
-
   if (payload.userEmail) {
-    const userExist = await models.User.findOne({ where: { email: payload.userEmail } });
-    if (!userExist) {
-      throw new Error('User Not Found');
-    }
-    console.log(payload);
-    return resetPassword(payload.newPassword,null,payload.userEmail);
-  }
-  else if (payload.token) {
-    if (!await get(payload.token)) {
-      throw new Error("reset link expired");
-    }
-    const decode_token = jwt.verify(payload.token, process.env.secretKey);
-    if (!decode_token) {
-      throw new Error('Invalid reset link');
-    }
-    const userExist = await models.User.findOne({ where: { id: decode_token.userId } });
-    if (!userExist) {
-      throw new Error('User Not Found');
-    }
-    await redis.del(payload.token);
-    return resetPassword(payload.newPassword,userExist.id, userExist.email);
-  }
-}
-
-const userInfo = async (payload) => {
-  let key_name = payload.userId;
-  let is_data = await redis.get(key_name);
-
-  if (is_data) {
-    return JSON.parse(is_data);
-  }
-  else {
-    const userInfo = await models.User.findOne({
-    where: { id: payload.userId },
-    include: [{
-      model: models.Role,
-      required: false,
-      attributes: ["role_title"]
-    }, {
-      model: models.Designation,
-      attributes: ["designation_title"]
-    }],
-    attributes: { exclude: ["password", "created_at", "updated_at", "deleted_at"] }
+    const userExist = await models.User.findOne({
+      where: { email: payload.userEmail },
+      include: { model: models.Role }
     });
-    await redis.set(key_name, JSON.stringify(userInfo));
-    return userInfo;
+    
+    if (!userExist) {
+      throw new Error('User Not Found');
+    }
+
+    if (payload.roleCode || userExist.Roles[0].role_title=='User') {
+      return resetPassword(payload.newPassword, userExist.email);
+    }
+    else{
+      throw new Error('Access Denied');
+    }
+
+  }
+  else if (payload.token) { 
+    const isData = await redisClient.get(payload.token);
+    if (!isData) {
+      throw new Error("Invalid Reset Link");
+    }
+    const userExist = await models.User.findOne({ where: { id:isData } });
+    if (!userExist) {
+      throw new Error('User Not Found');
+    }
+    await redisClient.del(payload.token);
+    return resetPassword(payload.newPassword, userExist.email);
   }
 }
 
 const userDetail = async (payload) => {
-  const user_id = payload.user_id;
-  let key_name = user_id;
-  const is_data = await redis.get(key_name);
+  let keyName = payload.userId;
+  const isData = await redisClient.get(keyName);
 
-  if (is_data) {
-    return JSON.parse(is_data);
+  if (isData) {
+    return JSON.parse(isData);
   }
   else {
-    const userDesignationData = await models.User.findOne({
+    const userData = await models.User.findOne({
       where: {
-        id: user_id
+        id: payload.userId
       },
       include: [{
         model: models.Designation,
@@ -181,8 +162,12 @@ const userDetail = async (payload) => {
       }],
       attributes: { exclude: ["password", "created_at", "updated_at", "deleted_at"] }
     });
-    await redis.set(key_name, JSON.stringify(userDesignationData));
-    return userDesignationData;
+
+    if (!userData) {
+      throw new Error("User Not Found");
+    }
+    await redisClient.set(keyName, JSON.stringify(userData));
+    return userData;
   }
 }
 
@@ -197,23 +182,17 @@ const forgetPassword = async (payload) => {
 
   if (!user) {
     throw new Error('User Not Found!');
-  }
-
-  let tokenData = {
-    userId: user.dataValues.id,
-    time: Date.now()
-  }
-  let signUserToken = jwt.sign(tokenData, process.env.secretKey, {
-    expiresIn: process.env.FPT_EXPIRES_IN
-  });
+  }   
+  let randomToken = UniqueStringGenerator.UniqueString();
+  let userId = user.dataValues.id;
   let baseUrl = process.env.BASE_URL;
-  let resetPassawordLink = `${baseUrl}/api/user/reset-password/${signUserToken}`;
+  let resetPassawordLink = `${baseUrl}/api/user/reset-password/${randomToken}`;
 
-  await redis.set(signUserToken, resetPassawordLink);
+  await redisClient.set(randomToken, userId);
+
   let recipient = email;
   let subject = "Reset Password Link";
   let body = `Password Reset Link:- ${resetPassawordLink}`;
-
 
   await mailer.sendMail(body, subject, recipient);
   return "send reset password link successfully";
@@ -222,30 +201,42 @@ const forgetPassword = async (payload) => {
 
 const deactivateUser = async (payload) => {
 
-  let { userId } = payload;
+  let { user_id } = payload;
+  let keyName = user_id;
   const user = await models.User.findOne({
     where: {
-      id: userId
-    }
+      id: user_id
+    },
+    include: [{
+      model: models.Role,
+      attributes: ["role_code"]
+    }],
+
   });
 
   if (!user) {
     throw new Error("User Not Found")
   }
 
-  let destroyUser = await models.User.destroy({
+   if (user.Roles[0].role_code == 1001) {
+     throw new Error("Access denied")
+  }
+
+   await models.User.destroy({
     where: {
-      id: userId
+      id: user_id
     }
   });
-  return "User Deactiveted"
+
+  await redisClient.del(keyName);
+  return "User deactivate";
 }
 
 const enableUser = async (payload) => {
-  let { userId } = payload;
+  let { user_id } = payload;
   let restoreUser = await models.User.restore({
     where: {
-      id: userId
+      id: user_id
     }
   });
   if (restoreUser) {
@@ -258,7 +249,7 @@ const enableUser = async (payload) => {
 
 const createUser = async (payload) => {
   payload.is_firsttime = true;
-  payload.password = await hash(payload.password, 10)
+  payload.password = await bcrypt.hash(payload.password, 10)
   const trans = await sequelize.transaction();
   try {
     const existingUser = await models.User.findOne({
@@ -340,7 +331,8 @@ const createUser = async (payload) => {
 
 const registration = async (payload) => {
   payload.is_firsttime = false;
-  payload.password = await hash(payload.password, 10);
+  payload.role_key = "USR";
+  payload.password = await bcrypt.hash(payload.password, 10);
   const existingUser = await models.User.findOne({
     where: { email: payload.email },
   });
@@ -348,15 +340,31 @@ const registration = async (payload) => {
     throw new Error("User already exists");
   }
   const user = await models.User.create(payload);
+  const userId = user.dataValues.id;
+  const role = await models.Role.findOne({
+    where: {
+      role_key: payload.role_key,
+    },
+  });
+  await models.UserRoleMapping.create({
+    user_id: userId,
+    role_id: role.id,
+  });
+
+  delete user.dataValues.password;
+  delete user.dataValues.created_at;
+  delete user.dataValues.updated_at;
+  delete user.dataValues.deleted_at;
   return user;
-}
+};
+
+
 module.exports = {
   loginUser,
   getAllUsers,
   refreshToken,
   logoutUser,
   resetUserPassword,
-  userInfo,
   userDetail,
   forgetPassword,
   deactivateUser,
